python-onvif-server/requirements.txt
PyYAML==6.0.1
netifaces==0.11.0
onvif-zeep==0.2.12
xmltodict==0.13.0